Database: Updatet stat-Table for Europe-Data, Replaced boundingboxes-Index
boundingboxindex war a UNIQUE INDEX, but sqlite was only able to use one column of it. Now we use sqlites R*Tree-Capability, which is made exactly to do what we need: Cope with data with more than one dimension and find the data that fits the boundinbox.

More details at https://sqlite.org/rtree.html