Added lookupcli
lookupcli gets coordinates in latitude and longitude and calls the code to find out which the lowest matching border is. It should check this by doing polygon-checks, but this is not implemented yet. This version just uses the boundingboxes to find the lowest border that could match. This could be good enough for many points but can fail for some.