changed lookup to the correct behavior
I'm enough confident in lookup to change it to the final behavior: it only looks at borders with adminlevel 8 and lower because I do not want to know about where in the city I am, but only in what city. Now lookup cancels the search on first result because I'm only interested in the first result. Searching all possible results in the database is still possible by defining DEBUGLOOKUP