finally fixed 'memory is not freed after pas 3'-bug
Everything I did before to return the memory to the OS did not realy work. The memory was returned to libc correctly, but libc did not return it to the OS.

This is because the heep only has a length and can only be freed up to the last allocated block. Even after freeing all the protobuf-stuff the memory was not returned to the OS because one block of some other memory, for example sqlite, was later in the heep thus blocking everything before from being returned to the OS.

The solution is to give all memory back to libc by shuting down all protobufs and the sqlite-memory and than forcing libc to return memory.