Added compiletime-option for infinite GeometryCache
The define INFINITECACHE disables tracking of the ussage of cache-elements and purging of unused geometrys. The GeometryCache never forgets a Geometry. In my tests this brings a minimal speedup (1m18.322s instead of 1m22.880s), but it is almost guaranteed to have a much higher memory consumtion wich is not limited by anything, so this option is default disabled